requirements.txt
# Core RAG & Orchestration Dependencies
langchain
langchain-community
sentence-transformers
faiss-cpu  # <--- Using FAISS (CPU version) instead of ChromaDB
python-dotenv # For .env file
google-generativeai # Or the client for your chosen LLM API
# numpy # Try without pinning first

# Document Loading
pypdf

# Optional but useful
ipykernel # For notebooks if you use them
# requests